popup-fenix: change the order that buttons appear
Change the order that some tools and stats appear/disappear when using the more/less buttons in the popup.

The per-host toggle settings are useful on their own without having the stats display or the global options bar.

This keeps the popup small while being able to access the per-host settings.